push

+++ controllers/tag.js
@@ -42,6 +42,10 @@ export const updateTag = async (req, res) => {
 
 export const deleteTag = async (req, res) => {
   const { tagId } = req.params;
+  const user = User.findOne({ _id: userId, admin: true });
+  if (!user) {
+    throw new NotFoundError(`User with id ${userId} who is an admin does not exist`);
+  }
   const tag = await Tag.findOneAndDelete({ _id: tagId });
   if (!tag) {
     throw new NotFoundError(`Tag with id ${tagId} does not exist`);
